requirements.txt
aniso8601==6.0.0
dj-database-url==0.5.0
Django==2.2.3
django-filter==2.1.0
django-graphql-jwt==0.2.1
graphene==2.1.6
graphene-django==2.3.2
graphql-core==2.2
graphql-relay==0.4.5
promise==2.2.1
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
